services/loader/requirements.txt
psycopg2-binary==2.9.9
geopandas==0.14.4
shapely==2.0.4
fiona==1.9.6
pandas==2.2.0
httpx==0.28.0
redis==5.2.0
click==8.1.7
ckanapi==4.8
